### 2024-07-01 #6947 @editorialbot
[PRE REVIEW]: gotranx - General ODE translator
**Submitting author:** @finsberg (<a href="http://orcid.org/0000-0003-3766-2393">Henrik Nicolay Topnes Finsberg</a>)
**Repository:** https://github.com/finsberg/gotranx
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
