### 2024-07-01 #6947 @finsberg
> 👋 @finsberg - I don't see an open source license in your repo, or discussion of one in your README. JOSS requires an OSI-approved license for software that we review, so if you can add one, we can proceed, and if not, we will mark the submission as rejected.

My apologies. I have now added MIT license to the repo
