### 2024-07-01 #6947 @danielskatz
Thanks. I'll next look for an editor. While I do so, I think the link in your readme to the contributing guide is broken, and you also might want to mention the license in your readme.
