### 2024-07-01 5c30427
Fix broken link in README and add section with license in README
