### 2024-07-01 8081f04
Update contributing guide with info about jupyter book and add placeholder for more docs
