+++ docs/about.md
@@ -1,0 +1,5 @@
+# About
+
+In this document we will try to explain the process that is happening behind the scenes when an `.ode` file is parsed and code is generated.
+
+More info will come later
