### 2024-07-02 612fa6e
Add spell checker and fix some spelling mistakes
