+++ examples/lorentz.py
@@ -1,0 +1,218 @@
+# # Your first ODE file
+# In this tutorial we will show how to write your own ODE file from scratch. We will use Lorentz system as an example. The Lorentz system is a system of ordinary differential equations first studied by Edward Lorenz. It is notable for having chaotic solutions for certain parameter values and initial conditions. In particular, the Lorenz attractor is a set of chaotic solutions of the Lorenz system which, when plotted, resemble a butterfly or figure eight.
+# The equations are given by:
+#
+# ```{math}
+# \begin{align}
+# \frac{dx}{dt} &= \sigma(y - x) \\
+# \frac{dy}{dt} &= x(\rho - z) - y \\
+# \frac{dz}{dt} &= x y - \beta z
+# \end{align}
+# ```
+# where $\sigma$, $\rho$ and $\beta$ are the parameters of the system.
+#
+# When solving the system we will use the following parameters:
+#
+# ```{math}
+# \begin{align}
+# \sigma &= 12 \\
+# \rho &= 21 \\
+# \beta &= 2.4
+# \end{align}
+# ```
+#
+# We will also use the following initial conditions:
+#
+# ```{math}
+# \begin{align}
+# x(0) &= 1.0 \\
+# y(0) &= 2.0 \\
+# z(0) &= 3.05
+# \end{align}
+# ```
+#
+# Let's start by writing the ODE file.
+
+ode_str = """
+# This is the Lorentz system
+# And it is part of this tutorial
+
+parameters(
+sigma=12.0,
+rho=21.0,
+beta=2.4
+)
+
+states(x=ScalarParam(1.0, unit="m", description="x variable"), y=2.0,z=3.05)
+
+dx_dt = sigma * (y - x)  # The derivative of x
+dy_dt = x * (rho - z) - y  # m/s
+dz_dt = x * y - beta * z
+"""
+
+# Note that we have also add a description and a unit to the state `x`, and some description and units to `dx_dt` and `dy_dt` respectively. We can now save the file to disk in a file called `lorentz.ode`
+
+# +
+from pathlib import Path
+
+Path("lorentz.ode").write_text(ode_str)
+# -
+
+# We can now load the ODE with `gotranx`
+#
+
+# +
+import gotranx
+
+ode = gotranx.load_ode("lorentz.ode")
+# -
+
+# Let us first print the ODE
+
+print(ode)
+
+# We can also print the states
+
+from pprint import pprint
+
+pprint(ode.states)
+
+# and we see that we have three states. Similarly we can print the parameters
+
+pprint(ode.parameters)
+
+# and the state derivatives.
+
+pprint(ode.state_derivatives)
+
+# We could also print the intermediates, but there are no intermediate variables (i.e variables that are not states, parameters nor state derivatives) in the model
+
+pprint(ode.intermediates)
+
+# In the ODE file we also added some text at the top. This can be accessed using the `text` attribute
+
+print(ode.text)
+
+# We can also get a dictionary with all the sympy symbols using i the model
+
+print(ode.symbols)
+
+# Here the keys are the names of the variables and the values are the sympy symbols. We can also see which variables depend on which variables
+
+print(ode.dependents())
+
+# Here for example `dz_dt` depends only on `beta`, while all state derivatives depend on the state `y`. Now let use take a closer look at one of the states, for example `x` where
+
+x = ode["x"]
+print(x)
+
+# Can print the value
+
+print(x.value)
+
+# which is actually a sympy object.
+
+print(type(x.value))
+
+# We can also print the unit
+
+print(x.unit)
+
+# which is a [pint](https://pint.readthedocs.io/en/stable/) unit object. The original string is stored in the `unit_str` attribute
+
[…]
