+++ src/gotranx/templates/jax.py
@@ -3,7 +3,7 @@ from textwrap import dedent, indent
 import functools
 from structlog import get_logger
 
-from .python import acc
+from .python import acc, state_index, parameter_index, monitor_index, missing_index
 
 logger = get_logger()
 
@@ -101,3 +101,14 @@ def {name}({args}):
 {indent_return}
 """,
     )
+
+
+__all__ = [
+    "init_state_values",
+    "init_parameter_values",
+    "method",
+    "parameter_index",
+    "state_index",
+    "monitor_index",
+    "missing_index",
+]
