### 2024-07-09 #6947 @HaoZeke
Thanks for the invite @danielskatz, I'm happy to accept

@finsberg, do you have any suggestions from the [reviewer database](https://reviewers.joss.theoj.org/lookup)? I will start sending out invitations by the end of this week.
