### 2024-07-09 #6947 @HaoZeke
Since this uses sympy under the hood...

@certik @asmeurer @isuruf 👋 would you be interested in and available to review this JOSS submission? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: [joss.readthedocs.io/en/latest/review_criteria.html](https://joss.readthedocs.io/en/latest/review_criteria.html)
