### 2024-07-10 1c8fc5f
Print out generated code at the end
