### 2024-07-10 88d39a4
Create two workflows for building docs
