+++ .github/workflows/build_docs.yml
@@ -1,0 +1,36 @@
+name: Build documentation
+
+on:
+  pull_request:
+    branches: [main]
+  workflow_call:
+  workflow_dispatch:
+
+
+jobs:
+
+  build:
+    runs-on: ubuntu-22.04
+    env:
+      PUBLISH_DIR: ./_build/html
+
+
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v4
+
+      - name: Setup Python
+        uses: actions/setup-python@v5
+        with:
+          python-version: "3.10"
+
+      - name: Install dependencies
+        run: python3 -m pip install ".[docs]"
+
+      - name: Build docs
+        run: jjupyter-book build -W --keep-going .
+
+      - name: Upload artifact
+        uses: actions/upload-pages-artifact@v3
+        with:
+          path: ${{ env.PUBLISH_DIR }}
