+++ .github/workflows/build_docs.yml
@@ -28,7 +28,7 @@ jobs:
         run: python3 -m pip install ".[docs]"
 
       - name: Build docs
-        run: jjupyter-book build -W --keep-going .
+        run: jupyter-book build -W --keep-going .
 
       - name: Upload artifact
         uses: actions/upload-pages-artifact@v3
