### 2024-07-10 8040a53
Write ode file directly in example
