### 2024-07-10 4c5a25a
Set default loglevel to info
