### 2024-07-11 7864bf1
Restructure scheme generation so that we can make calls to other functions
