+++ src/gotranx/schemes.py
@@ -1,4 +1,5 @@
 from __future__ import annotations
+from types import CodeType
 import typing
 import sympy
 from enum import Enum
@@ -8,6 +9,9 @@ from structlog import get_logger
 from . import atoms
 from .ode import ODE
 from . import sympytools
+
+if typing.TYPE_CHECKING:
+    from .codegen.base import RHSFunc
 
 logger = get_logger()
 
@@ -33,10 +37,13 @@ class printer_func(typing.Protocol):
 
 
 class scheme_func(typing.Protocol):
+    __code__: CodeType
+
     def __call__(
         self,
         ode: ODE,
         dt: sympy.Symbol,
+        rhs: RHSFunc,
         name: str = "values",
         printer: printer_func = default_printer,
         remove_unused: bool = False,
@@ -102,6 +109,7 @@ def fraction_numerator_is_nonzero(expr):
 def forward_explicit_euler(
     ode: ODE,
     dt: sympy.Symbol,
+    rhs: RHSFunc,
     name: str = "values",
     printer: printer_func = default_printer,
     remove_unused: bool = False,
@@ -154,6 +162,7 @@ def forward_explicit_euler(
 def forward_generalized_rush_larsen(
     ode: ODE,
     dt: sympy.Symbol,
+    rhs: RHSFunc,
     name: str = "values",
     printer: printer_func = default_printer,
     remove_unused: bool = False,
