+++ src/gotranx/cli/gotran2c.py
@@ -5,7 +5,7 @@ import structlog
 
 from ..codegen.c import CCodeGenerator
 from ..load import load_ode
-from ..schemes import Scheme
+from ..schemes import Scheme, get_scheme
 from ..ode import ODE
 
 logger = structlog.get_logger()
@@ -39,8 +39,7 @@ def get_code(
         ode, remove_unused=remove_unused, apply_clang_format=apply_clang_format
     )
     comp = [
-        "#include <math.h>",
-        "#include <string.h>\n",
+        codegen.imports(),
         f"int NUM_STATES = {len(ode.states)};",
         f"int NUM_PARAMS = {len(ode.parameters)};",
         f"int NUM_MONITORED = { len(ode.state_derivatives) + len(ode.intermediates)};",
@@ -54,7 +53,7 @@ def get_code(
     ]
     if scheme is not None:
         for s in scheme:
-            comp.append(codegen.scheme(s.value))
+            comp.append(codegen.scheme(get_scheme(s.value)))
 
     return codegen._format("\n".join(comp))
 
