+++ src/gotranx/cli/gotran2c.py
@@ -39,8 +39,7 @@ def get_code(
         ode, remove_unused=remove_unused, apply_clang_format=apply_clang_format
     )
     comp = [
-        "#include <math.h>",
-        "#include <string.h>\n",
+        codegen.imports(),
         f"int NUM_STATES = {len(ode.states)};",
         f"int NUM_PARAMS = {len(ode.parameters)};",
         f"int NUM_MONITORED = { len(ode.state_derivatives) + len(ode.intermediates)};",
