### 2024-07-11 783d8f2
Turn off interactive plotting when running tests
