### 2024-07-11 24b942f
Update about page. Add a few more checks when loading invalid ode and update exlude patterns
