### 2024-07-12 52f7c47
Make it possible pass in function to generate scheme rather than name
