+++ src/gotranx/cli/gotran2py.py
@@ -5,8 +5,10 @@ import structlog
 
 from ..codegen.python import PythonCodeGenerator
 from ..load import load_ode
-from ..schemes import Scheme, get_scheme
+from ..schemes import Scheme
 from ..ode import ODE
+
+from .utils import add_schemes
 
 logger = structlog.get_logger()
 
@@ -17,6 +19,8 @@ def get_code(
     apply_black: bool = True,
     remove_unused: bool = False,
     missing_values: dict[str, int] | None = None,
+    delta: float = 1e-8,
+    stiff_states: list[str] | None = None,
 ) -> str:
     """Generate the Python code for the ODE
 
@@ -30,6 +34,13 @@ def get_code(
         Apply black formatter, by default True
     remove_unused : bool, optional
         Remove unused variables, by default False
+    missing_values : dict[str, int] | None, optional
+        Missing values, by default None
+    delta : float, optional
+        Delta value for the rush larsen schemes, by default 1e-8
+    stiff_states : list[str] | None, optional
+        Stiff states, by default None. Only applicable for
+        the hybrid rush larsen scheme
 
     Returns
     -------
@@ -57,12 +68,12 @@ def get_code(
         codegen.rhs(),
         codegen.monitor_values(),
         _missing_values,
-    ]
-
-    if scheme is not None:
-        for s in scheme:
-            comp.append(codegen.scheme(get_scheme(s.value)))
-
+    ] + add_schemes(
+        codegen,
+        scheme=scheme,
+        delta=delta,
+        stiff_states=stiff_states,
+    )
     return codegen._format("\n".join(comp))
 
 
@@ -74,6 +85,8 @@ def main(
     scheme: list[Scheme] | None = None,
     remove_unused: bool = False,
     verbose: bool = True,
+    stiff_states: list[str] | None = None,
+    delta: float = 1e-8,
 ) -> None:
     loglevel = logging.DEBUG if verbose else logging.INFO
     structlog.configure(
