### 2024-07-12 70e6d26
Test that all schemes are generated
