### 2024-07-12 10a4ede
Try to use same versions as pydantic
