+++ pyproject.toml
@@ -60,7 +60,9 @@ test = [
     "jax",
 ]
 benchmark = [
-    "pytest-codspeed",
+    "pytest-benchmark>=4.0.0",
+    # TODO: remove bound once pytest-codspeed supports 3.13, blocked by `python-cffi`
+    "pytest-codspeed~=2.2.0; python_version < '3.13'",
     "gotranx[test]",
 ]
 
