+++ pyproject.toml
@@ -85,7 +85,6 @@ namespaces = false
 test = "pytest"
 
 [tool.pytest.ini_options]
-addopts = "--cov=gotranx --cov-report html --cov-report xml --cov-report term-missing -v"
 testpaths = ["tests"]
 
 [tool.isort]
