+++ .github/workflows/codspeed.yml
@@ -25,4 +25,4 @@ jobs:
         uses: CodSpeedHQ/action@v2
         with:
           token: ${{ secrets.CODSPEED_TOKEN }}
-          run: pytest tests/ --codspeed
+          run: pytest tests/test_benchmarks.py --codspeed
