### 2024-07-12 7bb6590
Try to add a few options for pytest-benchmark
