### 2024-07-12 0ded5d1
Create factory function for formatter
