### 2024-07-12 45e2746
Only apply black at the very end when running the cli
