### 2024-07-12 b75aa80
Use ruff instead of black
