### 2024-07-12 d9eba6b
Update README.md
Update badge for docs
