### 2024-07-13 3d666ee
Change name from 'apply_black' to 'format' and fix tests
