+++ pyproject.toml
@@ -83,8 +83,6 @@ test = "pytest"
 addopts = "--cov=gotranx --cov-report html --cov-report xml --cov-report term-missing -v"
 testpaths = ["tests"]
 
-[tool.isort]
-profile = "black"
 
 [tool.mypy]
 files = ["src/gotranx", "tests"]
