+++ src/gotranx/cli/gotran2py.py
@@ -3,7 +3,7 @@ from pathlib import Path
 import logging
 import structlog
 
-from ..codegen.python import PythonCodeGenerator, get_formatter
+from ..codegen.python import PythonCodeGenerator, get_formatter, Format
 from ..load import load_ode
 from ..schemes import Scheme
 from ..ode import ODE
@@ -16,7 +16,7 @@ logger = structlog.get_logger()
 def get_code(
     ode: ODE,
     scheme: list[Scheme] | None = None,
-    format: bool = True,
+    format: Format = Format.black,
     remove_unused: bool = False,
     missing_values: dict[str, int] | None = None,
     delta: float = 1e-8,
@@ -30,8 +30,8 @@ def get_code(
         The ODE
     scheme : list[Scheme] | None, optional
         Optional numerical scheme, by default None
-    format : bool, optional
-        Apply ruff / black formatter, by default True
+    format : Format, optional
+        The formatter, by default Format.black
     remove_unused : bool, optional
         Remove unused variables, by default False
     missing_values : dict[str, int] | None, optional
@@ -49,10 +49,10 @@ def get_code(
     """
     codegen = PythonCodeGenerator(
         ode,
-        format=False,
+        format=Format.none,
         remove_unused=remove_unused,
     )
-    formatter = get_formatter()
+    formatter = get_formatter(format=format)
     if missing_values is not None:
         _missing_values = codegen.missing_values(missing_values)
     else:
@@ -76,7 +76,9 @@ def get_code(
         stiff_states=stiff_states,
     )
     code = codegen._format("\n".join(comp))
-    if format:
+
+    if format != Format.none:
+        # Run the formatter only once
         code = formatter(code)
     return code
 
@@ -85,7 +87,7 @@ def main(
     fname: Path,
     suffix: str = ".py",
     outname: str | None = None,
-    format: bool = True,
+    format: Format = Format.black,
     scheme: list[Scheme] | None = None,
     remove_unused: bool = False,
     verbose: bool = True,
@@ -103,6 +105,7 @@ def main(
         scheme=scheme,
         format=format,
         remove_unused=remove_unused,
+        stiff_states=stiff_states,
         delta=delta,
     )
     out = fname if outname is None else Path(outname)
