+++ src/gotranx/codegen/__init__.py
@@ -3,8 +3,8 @@ from . import c
 from . import python
 from . import ode
 
-from .c import CCodeGenerator, GotranCCodePrinter
-from .python import PythonCodeGenerator, GotranPythonCodePrinter
+from .c import CCodeGenerator, GotranCCodePrinter, Format as CFormatter
+from .python import PythonCodeGenerator, GotranPythonCodePrinter, Format as PythonFormatter
 from .base import CodeGenerator, Func, RHSArgument, SchemeArgument
 from .ode import GotranODECodePrinter, BaseGotranODECodePrinter
 
@@ -23,4 +23,6 @@ __all__ = [
     "GotranODECodePrinter",
     "BaseGotranODECodePrinter",
     "GotranCCodePrinter",
+    "CFormatter",
+    "PythonFormatter",
 ]
