+++ src/gotranx/cli/__init__.py
@@ -1,5 +1,6 @@
 import typing
 from pathlib import Path
+import warnings
 
 try:
     from typing import Annotated
@@ -9,7 +10,8 @@ except ImportError:
 import typer
 
 from ..schemes import Scheme
-from . import gotran2c, gotran2py, cellml2ode
+from ..codegen import PythonFormatter, CFormatter
+from . import gotran2c, gotran2py
 
 app = typer.Typer()
 
@@ -49,6 +51,48 @@ def main(
         help="Show license",
     ),
 ): ...
+
+
+def find_pyproject_toml_config() -> Path | None:
+    """Find the pyproject.toml file."""
+    from black.files import find_pyproject_toml
+
+    path = find_pyproject_toml((str(Path.cwd()),))
+    if path is None:
+        return None
+    return Path(path)
+
+
+def read_config(path: Path | None) -> dict[str, typing.Any]:
+    """Read the configuration file."""
+
+    # If no path is given, try to find the pyproject.toml file
+    if path is None:
+        path = find_pyproject_toml_config()
+
+    # Return empty dict if no path is found
+    if path is None:
+        return {}
+
+    # Try to read the configuration file
+    try:
+        # First try to use tomllib which is part of stdlib
+        import tomllib as toml
+    except ImportError:
+        # If tomllib is not available, try to use toml
+        try:
+            import toml  # type: ignore
+        except ImportError:
+            typer.echo("Please install 'tomllib' or 'toml' to read configuration files")
+            return {}
+
+    try:
+        config = toml.loads(Path(path).read_text())
+    except Exception:
+        typer.echo(f"Could not read configuration file {path}")
+        return {}
+    else:
+        return config.get("tool", {}).get("gotranx", {})
 
 
 @app.command()
@@ -111,6 +155,12 @@ def convert(
         help="Delta value for the rush larsen schemes",
     ),
 ):
+    warnings.warn(
+        "convert command is deprecated, use ode2c, ode2py or cellml2ode instead",
+        DeprecationWarning,
+        stacklevel=1,
+    )
+
     if fname is None:
         return typer.echo("No file specified")
 
@@ -145,7 +195,237 @@ def convert(
         )
 
     if to in {".ode"}:
-        cellml2ode.main(fname=fname, outname=outname, verbose=verbose)
+        from .cellml2ode import main as _main
+
+        _main(fname=fname, outname=outname, verbose=verbose)
+
+
+@app.command()
+def cellml2ode(
+    fname: typing.Optional[Path] = typer.Argument(
+        None,
+        exists=True,
+        file_okay=True,
+        dir_okay=False,
+        writable=False,
+        readable=True,
+        resolve_path=True,
+    ),
+    outname: typing.Optional[str] = typer.Option(
+        None,
+        "-o",
+        "--outname",
+        help="Output name",
+    ),
+    version: bool = typer.Option(
+        None,
+        "--version",
+        callback=version_callback,
+        is_eager=True,
+        help="Show version",
+    ),
+    license: bool = typer.Option(
+        None,
+        "--license",
+        callback=license_callback,
+        is_eager=True,
+        help="Show license",
+    ),
+    config: typing.Optional[Path] = typer.Option(
+        None,
+        "--config",
+        help="Read configuration options from a configuration file",
+    ),
+    verbose: bool = typer.Option(
+        False,
+        "--verbose",
+        "-v",
+        help="Verbose output",
+    ),
+):
+    if fname is None:
+        return typer.echo("No file specified")
+
+    # config_data = read_config(config)
+    from .cellml2ode import main as _main
+
+    _main(fname=fname, outname=outname, verbose=verbose)
+
+
+@app.command()
+def ode2py(
[…]
