+++ src/gotranx/cli/__init__.py
@@ -1,3 +1,4 @@
+from __future__ import annotations
 import typing
 from pathlib import Path
 import warnings
