### 2024-07-14 ee5d9a7
Move cli utility funcitons to utils.py and remove __future__ import annotations
