+++ src/gotranx/cli/utils.py
@@ -1,5 +1,9 @@
 from __future__ import annotations
 from typing import Any
+from pathlib import Path
+
+import typer
+
 from ..codegen import CodeGenerator
 from ..schemes import Scheme, get_scheme
 
@@ -21,3 +25,45 @@ def add_schemes(
 
             comp.append(codegen.scheme(get_scheme(s.value), **kwargs))
     return comp
+
+
+def find_pyproject_toml_config() -> Path | None:
+    """Find the pyproject.toml file."""
+    from black.files import find_pyproject_toml
+
+    path = find_pyproject_toml((str(Path.cwd()),))
+    if path is None:
+        return None
+    return Path(path)
+
+
+def read_config(path: Path | None) -> dict[str, Any]:
+    """Read the configuration file."""
+
+    # If no path is given, try to find the pyproject.toml file
+    if path is None:
+        path = find_pyproject_toml_config()
+
+    # Return empty dict if no path is found
+    if path is None:
+        return {}
+
+    # Try to read the configuration file
+    try:
+        # First try to use tomllib which is part of stdlib
+        import tomllib as toml
+    except ImportError:
+        # If tomllib is not available, try to use toml
+        try:
+            import toml  # type: ignore
+        except ImportError:
+            typer.echo("Please install 'tomllib' or 'toml' to read configuration files")
+            return {}
+
+    try:
+        config = toml.loads(Path(path).read_text())
+    except Exception:
+        typer.echo(f"Could not read configuration file {path}")
+        return {}
+    else:
+        return config.get("tool", {}).get("gotranx", {})
