### 2024-07-14 9ba3e5a
Fix docstring in gotran2c
