+++ pyproject.toml
@@ -169,6 +169,11 @@ replace = 'version = "{new_version}"'
 
 [tool.gotranx]
 verbose = true
+delta = 1e-6
+scheme = [
+    "explicit_euler",
+    "generalized_rush_larsen",
+]
 
 [tool.gotranx.formatter]
 c = "clang-format"
