### 2024-07-15 a4950c7
Add some new tests
