### 2024-07-15 0c866c8
Add test for config
