### 2024-07-15 56ee762
Add docs about configuration and add command for listing schemes
