### 2024-07-15 545264a
Update name of forward_generalized_rush_lanrsen in cli.md
