### 2024-07-15 d1c60a4
Bump version: 1.0.0 → 1.1.0
