### 2024-07-22 961cfdd
[pre-commit.ci] pre-commit autoupdate
updates:
- [github.com/astral-sh/ruff-pre-commit: v0.5.2 → v0.5.4](https://github.com/astral-sh/ruff-pre-commit/compare/v0.5.2...v0.5.4)
- [github.com/pre-commit/mirrors-mypy: v1.10.1 → v1.11.0](https://github.com/pre-commit/mirrors-mypy/compare/v1.10.1...v1.11.0)
- [github.com/streetsidesoftware/cspell-cli: v8.10.1 → v8.11.0](https://github.com/streetsidesoftware/cspell-cli/compare/v8.10.1...v8.11.0)
