+++ .pre-commit-config.yaml
@@ -14,7 +14,7 @@ repos:
 
   - repo: https://github.com/astral-sh/ruff-pre-commit
     # Ruff version.
-    rev: 'v0.5.2'
+    rev: 'v0.5.4'
     hooks:
        # Run the linter.
       - id: ruff
@@ -23,7 +23,7 @@ repos:
       - id: ruff-format
 
   - repo: https://github.com/pre-commit/mirrors-mypy
-    rev: v1.10.1
+    rev: v1.11.0
     hooks:
       - id: mypy
         additional_dependencies: ["types-attrs"]
@@ -31,7 +31,7 @@ repos:
         args: ["--config-file", "pyproject.toml"]
 
   - repo: https://github.com/streetsidesoftware/cspell-cli
-    rev: v8.10.1
+    rev: v8.11.0
     hooks:
       - id: cspell
         files: src/|docs/(.+).md|tests|README.md|examples/(.+).py
