### 2024-07-29 #6947 @HaoZeke
hi  @anilbey @benlansdell @SunnyXu @ayush9pandey 👋 would you be interested in and available to review this JOSS submission? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: [joss.readthedocs.io/en/latest/review_criteria.html](https://joss.readthedocs.io/en/latest/review_criteria.html)

If not, could you recommend any potential reviewers?
