+++ examples/run-python/main.py
@@ -12,6 +12,7 @@ import matplotlib.pyplot as plt
 # We load the model using the {py:func}`load_ode` function
 
 ode = gotranx.load_ode("ORdmm_Land.ode")
+
 # Now we can generate code in python using the `cli` subpackage and the `gotran2py` module`. We will also generate code for the generalized rush larsen scheme
 
 code = gotranx.cli.gotran2py.get_code(
