### 2024-08-01 finsberg/gotranx#95 @SunnyXu
Data sharing-ORdmm_Land.ode file
Regarding to the requirement from JOSS about the General checks, “Data sharing” is required. The ORdmm_Land.ode file is not provided either under the subsection “Using the Python API” or “Splitting an ODE into two sub-ODEs” or “Compiling a C-extension”. I see it is under examples folder on Github. It would be good to point to the .ode file in the corresponding documentation.
