### 2024-08-01 finsberg/gotranx#96 @SunnyXu
Automated tests guide in documentation
Regarding to the requirement from JOSS about the Documentation, “Automated tests” are required. I see there is a tests folder on GitHub. It would be good to add a section regarding how to run the tests in the documentation.
