### 2024-08-01 finsberg/gotranx#97 @SunnyXu
 A statement of need in documentation
Regarding to the requirement from JOSS about the Documentation, “A statement of need” is required. Maybe it has been briefly mentioned in the documentation, however, please add a specific section for it in the documentation.
