### 2024-08-01 finsberg/gotranx#98 @SunnyXu
Quick start
It would be good to give a brief guide of steps in words for Quick Start, instead of just pointing it to the documentation.
