### 2024-08-01 finsberg/gotranx#99 @SunnyXu
Adding a new numerical scheme
The subsection “Adding a new numerical scheme” under the section “How to guides” is not completed.
