### 2024-08-01 finsberg/gotranx#101 @SunnyXu
citations format
The citations in the manuscript are sometimes weird, which should be names followed by years. Please correct the citation on Line 37, and Line 44.
