### 2024-08-02 finsberg/gotranx#102 @SunnyXu
acknowledge financial support
According to the JOSS paper format (https://joss.readthedocs.io/en/latest/paper.html#what-should-my-paper-contain) Please acknowledge any financial support.
