### 2024-08-02 #7063 @SunnyXu
@HaoZeke @finsberg

Gotranx is a solid python package as an ODE translator following Gotran, which is applicable to the field of computational biology, i.e., heart cells. The paper is written clearly and easy to follow. Previous packages and related articles are well referred to. However, please see my comments for possible improvements.

(1) Data sharing-ORdmm_Land.ode file
https://github.com/finsberg/gotranx/issues/95

(2) Automated tests guide in documentation
https://github.com/finsberg/gotranx/issues/96

(3) A statement of need in documentation
https://github.com/finsberg/gotranx/issues/97

(4) Quick start
https://github.com/finsberg/gotranx/issues/98

(5) Adding a new numerical scheme
https://github.com/finsberg/gotranx/issues/99

(6) citations format
https://github.com/finsberg/gotranx/issues/101

(7) financial support
https://github.com/finsberg/gotranx/issues/102
