### 2024-08-02 2eb0e9a
Remove section on adding new scheme
