+++ _toc.yml
@@ -12,7 +12,6 @@ parts:
   - caption: How to guides
     chapters:
     - file: "examples/new-language/main.py"
-    - file: "docs/new_scheme"
     - file: "docs/config"
     - file: "examples/split-ode/main.py"
     - file: "examples/compile-c-extension/main.py"
