### 2024-08-02 finsberg/gotranx#99 @finsberg
Adressed in https://github.com/finsberg/gotranx/pull/103
