+++ joss-paper/paper.md
@@ -43,9 +43,9 @@ Solving ODEs can be done analytically in simple cases but for most real world ap
 
 Computational biology is one field where ODEs play an essential role, for example in models of heart cells. The resulting ODEs can, in these cases, be quite involved with many parameters and state variables. For example one of the more recent models of human heart cells [@tomek2019development] have 112 parameters, 45 state variables and 276 intermediate variables. Solutions of the state equations of these membrane models typically allows for specialized exponential integrators [@rush1978practical] that are implemented as a numerical scheme in `gotranx`.
 
-Furthermore, such models of a single heart cell are typically embedded into a spatial model, such as the Monodomain or Bidomain model [@sundnes2007computing], where each point in a 3D geometry represents a cell. This means that thousands or even millions of such ODEs needs to be coupled and solved in a larger systems. In these cases, traditional solvers are usually unsuitable and custom code are often developed to solve the ODEs. An alternative to this approach is to use an existing framework specialized for these types of problems [@cooper2020chaste, @plank2021opencarp]. However, introducing a big framework might not be ideal if the user wants to do a lot of customizations. With `gotranx` you can easily generate framework independent code that can easily be integrated into most simulation pipelines. One example one example is [`fenics-beat`](https://github.com/finsberg/fenics-beat) which uses `gotranx` to generate code solving ODEs in a Monodomain model.
+Furthermore, such models of a single heart cell are typically embedded into a spatial model, such as the Monodomain or Bidomain model [@sundnes2007computing], where each point in a 3D geometry represents a cell. This means that thousands or even millions of such ODEs needs to be coupled and solved in a larger systems. In these cases, traditional solvers are usually unsuitable and custom code are often developed to solve the ODEs. An alternative to this approach is to use an existing framework specialized for these types of problems [@cooper2020chaste,@plank2021opencarp]. However, introducing a big framework might not be ideal if the user wants to do a lot of customizations. With `gotranx` you can easily generate framework independent code that can easily be integrated into most simulation pipelines. One example one example is [`fenics-beat`](https://github.com/finsberg/fenics-beat) which uses `gotranx` to generate code solving ODEs in a Monodomain model.
 
-While such a models are typically developed in one programming language, different research groups use different programming languages to integrate and solve their models. To make these models programming language independent, it is common practice to write them in a markup language [@lloyd2004cellml, keating2020sbml]. However, when translating the models to a new programming language, user typically need to do this manually which are likely to introduce errors in the code. Since `gotranx` also implements converters for e.g CellML through MyoKit [@clerx2016myokit], it already supports most models that are used today.
+While such a models are typically developed in one programming language, different research groups use different programming languages to integrate and solve their models. To make these models programming language independent, it is common practice to write them in a markup language [@lloyd2004cellml,@keating2020sbml]. However, when translating the models to a new programming language, user typically need to do this manually which are likely to introduce errors in the code. Since `gotranx` also implements converters for e.g CellML through MyoKit [@clerx2016myokit], it already supports most models that are used today.
 
 
 # Acknowledgements
