### 2024-08-02 518bf71
Only one citation per bracket
