### 2024-08-02 finsberg/gotranx#101 @finsberg
Fixed in https://github.com/finsberg/gotranx/pull/104
