### 2024-08-02 finsberg/gotranx#102 @finsberg
Fixed in https://github.com/finsberg/gotranx/pull/106
