### 2024-08-03 1f0c623
Update README.md
Add joss status badge
