+++ src/gotranx/templates/julia.py
@@ -73,17 +73,17 @@ def method_index(data: dict[str, int], method_name) -> str:
         """
     {if_stm} (name == "{name}")
         return {index}
-    end
     """
     )
     code = []
     code.append(f"# {method_name.capitalize()} index")
-    code.append(f"function {method_name}_index(const char name[])")
+    code.append(f"function {method_name}_index(name::String)")
     code.append("{")
     for i, (name, index) in enumerate(data.items()):
         if_stm = "if" if i == 0 else "elseif"
         code.append(indent(local_template.format(if_stm=if_stm, name=name, index=index), "    "))
-    code.append(indent("return -1;", "    "))
+    code.append(indent("end",  "    "))
+    code.append(indent("return -1", "    "))
     code.append("end")
 
     return "\n".join(code)
