+++ src/gotranx/cli/__init__.py
@@ -510,6 +510,7 @@ def ode2julia(
         delta=delta,
     )
 
+
 @app.command()
 def list_schemes():
     from rich.console import Console
