### 2024-08-06 83686a1
int -> const for globals
