### 2024-08-09 finsberg/gotranx#110 @ayush9pandey
JOSS Review: [Suggestion] examples/lorentz.py is better formatted as a jupyter notebook
The lorentz.py is a great example to use as a tutorial for the package. However, it contains multiple print statements to show various features of the tool. For such files, jupyter notebook is a better visual, in my opinion. It could replace the lorentz.py file or can be given as an alternate.
