### 2024-08-09 finsberg/gotranx#110 @finsberg
I agree that jupyter notebooks are good for this, and in the documentation (i.e https://finsberg.github.io/gotranx/examples/lorentz.html) the file is actually converted to a notebook using [`jupytext`](https://jupytext.readthedocs.io/en/latest/) before is i passed to JupyterBook to, see https://github.com/finsberg/gotranx/blob/main/_config.yml#L38-L41

I prefer to keep it as a python file, because it is much cleaner with version control.
