### 2024-08-10 0c8aafc
Add quick start and add initial value for functools.reduce in python template
