### 2024-08-10 898b537
Add quick start and add initial value for functools.reduce in python template
