+++ README.md
@@ -32,7 +32,54 @@ python3 -m pip install git+https://github.com/finsberg/gotranx
 ```
 
 ## Quick start
-Check out the demos in the [documentation](https://finsberg.github.io/gotranx/)
+Define your ODE in a `.ode` file, e.g `file.ode` with the current content
+```
+states(x=1, y=0)
+
+dx_dt = y
+dy_dt = -x
+```
+which defines the ode system
+$$
+\dot{x} = y
+\dot{y} = -x
+$$
+with the initial conditions $x(0) = 1$ and $y(0) = 0$. Now generate code in python for solving this ODE with the explicit euler scheme using the command
+```
+gotranx ode2py file.ode --scheme explicit_euler -o file.py
+```
+which will create a file `file.py` containing functions for solving the ODE. Now you can solve the ode using the the following code snippet
+
+```python
+import file as model
+import numpy as np
+import matplotlib.pyplot as plt
+
+s = model.init_state_values()
+p = model.init_parameter_values()
+dt = 1e-4  # 0.1 ms
+T = 2 * np.pi
+t = np.arange(0, T, dt)
+
+x_index = model.state_index("x")
+x = [s[x_index]]
+y_index = model.state_index("y")
+y = [s[y_index]]
+
+for ti in t[1:]:
+    s = model.explicit_euler(s, ti, dt, p)
+    x.append(s[x_index])
+    y.append(s[y_index])
+
+plt.plot(t, x, label="x")
+plt.plot(t, y, label="y")
+plt.legend()
+plt.show()
+```
+![_](docs/_static/quick_start.png)
+
+
+Note that this is a rather artificial example, where it is probably simpler to just use [`scipy.integrate.solve_ivp`](https://docs.scipy.org/doc/scipy/reference/generated/scipy.integrate.solve_ivp.html) instead, so check out the demos in the [documentation](https://finsberg.github.io/gotranx/) for more elaborate examples.
 
 
 ## License
