### 2024-08-10 finsberg/gotranx#112 @ayush9pandey
JOSS Review: Additional requirements for running all tests?
I'm trying to run the unit tests locally on my Windows computer but running into errors with all the CLI tests (and a few more). To attempt to fix these, I've installed LLVM and clang on my anaconda environment but this does not seem to help. I am struggling to find documentation (if there's any) on setting up a Python environment that can run all tests.  It seems that there's an issue with running CLI specific tests.

Error traceback:
```
================================================== short test summary info ==================================================
FAILED tests/test_benchmarks.py::test_c_no_schemes - subprocess.CalledProcessError: Command '['C:\\Users\\ayush\\anaconda3\\Library\\bin\\clang-format.EXE', '--style=Microsof...
FAILED tests/test_benchmarks.py::test_c_explicit_euler - subprocess.CalledProcessError: Command '['C:\\Users\\ayush\\anaconda3\\Library\\bin\\clang-format.EXE', '--style=Microsof...
FAILED tests/test_benchmarks.py::test_c_generalized_rush_larsen - subprocess.CalledProcessError: Command '['C:\\Users\\ayush\\anaconda3\\Library\\bin\\clang-format.EXE', '--style=Microsof...
FAILED tests/test_c_codegen.py::test_c_codegen_initial_state_values - AssertionError: assert '\nvoid init_... = 3.05;\n}\n' == '\nvoid init_... = 3.05;\n}\n'
FAILED tests/test_c_codegen.py::test_c_codegen_initial_parameter_values - AssertionError: assert '\nvoid init_... = 12.0;\n}\n' == '\nvoid init_... = 12.0;\n}\n'
FAILED tests/test_c_codegen.py::test_c_codegen_rhs[stp-const double *__restrict states, const double t, const double *__restrict parameters, double *values] - AssertionError: assert '\nvoid rhs(c...= dz_dt;\n}\n' == '\nvoid rhs(c...= dz_dt;\n}\n'
FAILED tests/test_c_codegen.py::test_c_codegen_rhs[spt-const double *__restrict states, const double *__restrict parameters, const double t, double *values] - AssertionError: assert '\nvoid rhs(c...= dz_dt;\n}\n' == '\nvoid rhs(c...= dz_dt;\n}\n'
FAILED tests/test_c_codegen.py::test_c_codegen_rhs[tsp-const double t, const double *__restrict states, const double *__restrict parameters, double *values] - AssertionError: assert '\nvoid rhs(c...= dz_dt;\n}\n' == '\nvoid rhs(c...= dz_dt;\n}\n'
FAILED tests/test_c_codegen.py::test_c_codegen_forward_euler - AssertionError: assert '\nvoid forwa..._dt + z;\n}\n' == '\nvoid forwa..._dt + z;\n}\n'
FAILED tests/test_c_codegen.py::test_c_codegen_forward_generalized_rush_larsen - AssertionError: assert '\nvoid forwa...n    ));\n}\n' == '\nvoid forwa...dz_dt));\n}\n'
FAILED tests/test_c_codegen.py::test_c_remove_unused_rhs - AssertionError: assert '\nvoid rhs(c...= dz_dt;\n}\n' == '\nvoid rhs(c...= dz_dt;\n}\n'
FAILED tests/test_c_codegen.py::test_c_remove_unused_forward_explicit_euler - AssertionError: assert '\nvoid forwa..._dt + z;\n}\n' == '\nvoid forwa..._dt + z;\n}\n'
FAILED tests/test_c_codegen.py::test_c_remove_unused_forward_generalized_rush_larsen - AssertionError: assert '\nvoid forwa...n    ));\n}\n' == '\nvoid forwa...dz_dt));\n}\n'
FAILED tests/test_c_codegen.py::test_c_conditional_expression_assignment - AssertionError: assert '\nvoid rhs(c...))/6.8);\n}\n' == '\nvoid rhs(c... / 6.8);\n}\n'
FAILED tests/test_c_codegen.py::test_c_conditional_expression_advanced - AssertionError: assert '\nvoid rhs(c...n    ));\n}\n' == '\nvoid rhs(c... (0))));\n}\n'
FAILED tests/test_c_codegen.py::test_c_codegen_parameter_index - assert '// Parameter...return -1;\n}' == '// Parameter...return -1;\n}'
FAILED tests/test_c_codegen.py::test_c_codegen_state_index - assert '// State ind...return -1;\n}' == '// State ind...return -1;\n}'
FAILED tests/test_c_codegen.py::test_c_codegen_monitor_index - assert '// Monitor i...return -1;\n}' == '// Monitor i...return -1;\n}'
FAILED tests/test_c_codegen.py::test_c_codegen_monitor - AssertionError: assert '\nvoid monit...= dz_dt;\n}\n' == '\nvoid monit...= dz_dt;\n}\n'
FAILED tests/test_cli.py::test_gotran2c_old - AssertionError: assert 1 == 0
FAILED tests/test_cli.py::test_gotran2c[.h-clang-format] - AssertionError: assert 1 == 0
FAILED tests/test_cli.py::test_gotran2c[.h-none] - AssertionError: assert 1 == 0
FAILED tests/test_cli.py::test_gotran2c[.c-clang-format] - AssertionError: assert 1 == 0
FAILED tests/test_cli.py::test_gotran2c[.c-none] - AssertionError: assert 1 == 0
FAILED tests/test_cli.py::test_ode2c_config_file - AssertionError: assert 1 == 0
FAILED tests/test_demos.py::test_examples[compile-c-extension] - subprocess.CalledProcessError: Command '['C:\\Users\\ayush\\anaconda3\\python.exe', 'main.py']' returned non-zero exit st...
FAILED tests/test_demos.py::test_examples[new-language] - subprocess.CalledProcessError: Command '['C:\\Users\\ayush\\anaconda3\\python.exe', 'main.py']' returned non-zero exit st...
```
