### 2024-08-10 5b28cc2
Fix tests and add test when no parameters are present in the ODE
