### 2024-08-10 7efaf05
Do not use clang-format-docs on windows
