+++ tests/test_c_codegen.py
@@ -34,6 +34,7 @@ def codegen(ode) -> CCodeGenerator:
     return CCodeGenerator(ode)
 
 
+@pytest.mark.skipif(sys.platform == "win32", reason="clang-format-docs is not available on Windows")
 def test_c_codegen_initial_state_values(codegen: CCodeGenerator):
     assert codegen.initial_state_values() == (
         "\nvoid init_state_values(double *states)\n"
@@ -48,6 +49,7 @@ def test_c_codegen_initial_state_values(codegen: CCodeGenerator):
     )
 
 
+@pytest.mark.skipif(sys.platform == "win32", reason="clang-format-docs is not available on Windows")
 def test_c_codegen_initial_parameter_values(codegen: CCodeGenerator):
     assert codegen.initial_parameter_values() == (
         "\nvoid init_parameter_values(double *parameters)\n"
@@ -79,6 +81,7 @@ def test_c_codegen_initial_parameter_values_no_clang_format(ode):
         )
 
 
+@pytest.mark.skipif(sys.platform == "win32", reason="clang-format-docs is not available on Windows")
 @pytest.mark.parametrize(
     "order, arguments",
     [
@@ -132,6 +135,7 @@ def test_c_codegen_rhs(order: str, arguments: str, codegen: CCodeGenerator):
     )
 
 
+@pytest.mark.skipif(sys.platform == "win32", reason="clang-format-docs is not available on Windows")
 def test_c_codegen_forward_euler(codegen: CCodeGenerator):
     assert codegen.scheme(get_scheme("forward_euler")) == (
         "\nvoid forward_euler(const double *__restrict states, const double t, const double dt,"
@@ -161,6 +165,7 @@ def test_c_codegen_forward_euler(codegen: CCodeGenerator):
     )
 
 
+@pytest.mark.skipif(sys.platform == "win32", reason="clang-format-docs is not available on Windows")
 def test_c_codegen_forward_generalized_rush_larsen(codegen: CCodeGenerator):
     assert codegen.scheme(get_scheme("forward_generalized_rush_larsen")) == (
         "\nvoid forward_generalized_rush_larsen(const double *__restrict states, const double t, const double dt,"
@@ -196,6 +201,7 @@ def test_c_codegen_forward_generalized_rush_larsen(codegen: CCodeGenerator):
     )
 
 
+@pytest.mark.skipif(sys.platform == "win32", reason="clang-format-docs is not available on Windows")
 def test_c_remove_unused_rhs(ode_unused):
     codegen_orig = CCodeGenerator(ode_unused)
     assert codegen_orig.rhs() == (
@@ -257,6 +263,7 @@ def test_c_remove_unused_rhs(ode_unused):
     )
 
 
+@pytest.mark.skipif(sys.platform == "win32", reason="clang-format-docs is not available on Windows")
 def test_c_remove_unused_forward_explicit_euler(ode_unused):
     codegen_orig = CCodeGenerator(ode_unused)
     assert codegen_orig.scheme(get_scheme("forward_euler")) == (
@@ -321,6 +328,7 @@ def test_c_remove_unused_forward_explicit_euler(ode_unused):
     )
 
 
+@pytest.mark.skipif(sys.platform == "win32", reason="clang-format-docs is not available on Windows")
 def test_c_remove_unused_forward_generalized_rush_larsen(ode_unused):
     codegen_orig = CCodeGenerator(ode_unused)
     assert codegen_orig.scheme(get_scheme("forward_generalized_rush_larsen")) == (
@@ -397,6 +405,7 @@ def test_c_remove_unused_forward_generalized_rush_larsen(ode_unused):
     )
 
 
[…]
