+++ .github/workflows/main.yml
@@ -11,9 +11,11 @@ jobs:
     name: Test on ${{ matrix.os }} with Python ${{ matrix.python-version }}
     runs-on: ${{ matrix.os }}
     strategy:
+      fail-fast: false
       matrix:
         python-version: ["3.8", "3.9", "3.10", "3.11"]
         os: [ubuntu-latest, windows-latest, macos-latest]
+
 
     steps:
     - uses: actions/checkout@v4
