### 2024-08-10 finsberg/gotranx#112 @finsberg
Thanks for pointing this out! I only tested it in ubuntu, but now that I am adding tests on windows as well (here: https://github.com/finsberg/gotranx/pull/113) I see the same issue. This is stems from an issue in `clang-format-docs` (see https://github.com/finsberg/clang-format-docs/issues/2), so my solution to this would be to turn off clang-format if you are using windows.

Again, thanks for spotting this one!
