### 2024-08-11 7df54f6
Use cmake to build rather than make in an attempt to make it work on windows
