+++ examples/compile-c-extension/utils.py
@@ -53,8 +53,8 @@ def build_c(model):
 
     build_dir.mkdir()
 
-    sp.check_call(["cmake", f"-DCELL_LIBFILE={model_name}", ".."], cwd=build_dir)
-    sp.check_call(["make"], cwd=build_dir)
+    sp.check_call(["cmake", "-S", ".", "-B", str(build_dir), f"-DCELL_LIBFILE={model_name}"])
+    sp.check_call(["cmake", "--build", str(build_dir)])
 
 
 def gotran2c(odefile):
