### 2024-08-11 efa26ca
Use \frac rather than \dot
