### 2024-08-11 c28b985
Use \frac rather than \dot
