### 2024-08-11 finsberg/gotranx#98 @finsberg
Fixed in https://github.com/finsberg/gotranx/pull/111
