### 2024-08-11 finsberg/gotranx#112 @finsberg
Addressed in https://github.com/finsberg/gotranx/pull/113
