### 2024-08-11 finsberg/gotranx#112 @ayush9pandey
Thanks! All tests that failed earlier pass now! I do have a failing test in `test_demos.py` due to missing `jax`. On installing it, all tests pass. It is not listed as a dependency in `pyproject.toml`, maybe adding it will help.
