### 2024-08-11 finsberg/gotranx#115 @ayush9pandey
JOSS Review: Issues with tutorials
When running through the [Using the Python API](https://finsberg.github.io/gotranx/examples/run-python/main.html) and the [Your first ODE file](https://finsberg.github.io/gotranx/examples/lorentz.html) page, I am running into a few issues:

1. I don't see a link to the `ORdmm_Land.ode` file. I was able to find it eventually under examples/split-ode/, so it might be better to mention the location on the documentation page.
2. As a user not familiar with cardiac physiology models, it's difficult to understand the model filename or what it means. This tutorial is placed before the "Your first ODE file", so at this point it is difficult to understand .ode files for new users.
3. Not clear what's printed when I save an ODE file (the number 294).
![image](https://github.com/user-attachments/assets/b32ddf32-3317-431c-92ea-5ba8469e3fe1)
4. On similar lines as the points above, it might be helpful if there was a tutorial on integrating other ode solver schemes. Maybe this is already given somewhere..
