### 2024-08-11 6efdee9
Add info about how to run tests
