+++ .cspell_dict.txt
@@ -9,6 +9,7 @@ cfile
 cmodel
 codegen
 codeprinter
+codspeed
 conds
 cpath
 ctypeslib
@@ -63,6 +64,7 @@ ptsd
 pycode
 pyplot
 pyproject
+pytest
 Radau
 Rdmm
 relop
