### 2024-08-11 33468eb
Add info about how to run tests
