### 2024-08-11 finsberg/gotranx#112 @finsberg
Good point! You should install the test dependencies, i.e
```
python3 -m pip install gotranx[test]
```
I have updated the REAMDE with instructions in https://github.com/finsberg/gotranx/issues/96
