+++ .cspell_dict.txt
@@ -43,6 +43,7 @@ linalg
 linearization
 linestyle
 linspace
+literalinclude
 logicalfunc
 matplotlib
 MPLBACKEND
