+++ examples/compile-c-extension/main.py
@@ -5,18 +5,22 @@
 # To start with we will just run through an example how how this can be used.
 #
 # First we import the modules `utils` where we add all the functionality as well as `matplotlib` for plotting
-
+#
+# ```{note}
+# The full source code for the all the files need for this demo (including `utils.py`) is found at the bottom of this document
+# ```
+#
 
 import matplotlib.pyplot as plt
 import utils
 
+
+# For this tutorial we will use a rather large system of ODE which simulated the electromechanics in cardiac cells that are based on the [O'Hara-Rudy model for electrophysiology](https://journals.plos.org/ploscompbiol/article?id=10.1371/journal.pcbi.1002061) and the [Land model](https://www.sciencedirect.com/science/article/abs/pii/S0022282817300639). You can download the model in `.ode` format {download}`here <./ORdmm_Land.ode>`
 #
 # Next we load the model. This is function contains the functionality for generating code and compiling the C-extension. Currently it will also regenerate the code as well as recompiling the code code every time you run the code. It is also possible to only do this if the relevant files do not exist.
-#
 
 model = utils.load_model("ORdmm_Land.ode", rebuild=True, regenerate=True)
 
-#
 # Next we get the initial states and parameters. The parameters we get back from `initial_parameter_values` are a `numpy` array. We also convert the parameters to a dictionary which is easier to work with since we can use the name of the parameter rather than the index.
 
 y = model.initial_state_values()
@@ -56,3 +60,21 @@ ax[0, 1].set_ylabel("Ca (mM)")
 ax[1, 1].set_ylabel("Istim (uA/cm^2)")
 fig.tight_layout()
 plt.show()
+
+# ## Source code
+#
+# ### `utils.py`
+# ```{literalinclude} ./utils.py
+# ```
+#
+# ### `cmodel.py`
+# ```{literalinclude} ./cmodel.py
+# ```
+#
+# ### `template.c`
+# ```{literalinclude} ./template.c
+# ```
+#
+# ### `CMakeLists.txt`
+# ```{literalinclude} ./CMakeLists.txt
+# ```
