+++ examples/compile-c-extension/CMakeLists.txt
@@ -16,6 +16,5 @@ set(CMAKE_C_FLAGS_RELEASE "-O3")
 set(CMAKE_RUNTIME_OUTPUT_DIRECTORY ${PROJECT_BINARY_DIR}/bin)
 set(CMAKE_LIBRARY_OUTPUT_DIRECTORY ${PROJECT_BINARY_DIR}/lib)
 
-# add_executable(${EXECUTABLE_OUTPUT_PATH} demo.c)
 add_library(${CELL_LIBFILE} SHARED demo.c)
 install(TARGETS ${CELL_LIBFILE} DESTINATION lib)
