### 2024-08-11 965e823
Add scipy.integrate.solve_ivp example in README
