### 2024-08-11 5b9686d
Add scipy.integrate.solve_ivp example in README
