### 2024-08-11 finsberg/gotranx#115 @finsberg
Thanks for your comments. I have now addressed all of these points here: https://github.com/finsberg/gotranx/pull/117
The only exception is point 3. 294 refers to the number of characters written to the file which is the return value from `Path.write_text`. I see two possible ways to fix this
1. Assign the value value to a variable, i.e
```python
num_chars = Path("lorentz.ode").write_text(ode_str)
```
which now introduces an unused variable. Alternatively, I could assign it to `_`, i.e
```python
_ = Path("lorentz.ode").write_text(ode_str)
```
but I also think this looks a bit weired.
2. I could write a sentence saying that 294 is just the number of characters. However, I think this also sounds a bit too much.

Do you have any opinions here?
