### 2024-08-12 finsberg/gotranx#115 @ayush9pandey
I agree -- this is not a major issue either way. My approach in the past has been to create a package specific `write_model` function, which in your case could be `write_ode(ode_string, filename)`. That way, you can wrap around the Path library's function and get the desired behavior.
