### 2024-08-12 #7063 @ayush9pandey
@HaoZeke Gotranx is a library for translating ODE from a high-level string-based language into executable code (Python and C code generation is currently supported). API documentation to add other languages for code generation is provided with an example of including `jax` code gen with gotranx for enhanced ODE integration performance. Examples from computational physiology are provided with the package but the features of ODE translation are agnostic to the application and can be broadly applied to any scientific discipline where ODEs are used for modeling. I have pointed out a few issues to the authors that they may consider in improving this package.

Comments on the software, as Github issues:
1. https://github.com/finsberg/gotranx/issues/115
2. https://github.com/finsberg/gotranx/issues/112
3. https://github.com/finsberg/gotranx/issues/110
4. Many of the computational physiology related integrations are provided, but I couldn't find the "vanilla" ODE integrator (scipy/odeint). It is probably not suited for large model applications but still worth adding the compatibility to welcome users from more areas.

Comments on the paper:
1. The motivation for developing a full rewrite of gotran is not clear. Why did the authors undertake the endeavor of rewriting gotran? This is perhaps suited for discussion in the statement of need.
2. I would recommend including a short description/summary of features of gotranx in the paper. I believe it would still be under the word limit if that section was included.
3. The authors might also consider including an illustrative graphic in the paper. They already have a great illustration on the readme file, maybe it can be adapted to a particular example of simple ODE and included in the paper.
4. Plank et al. citation format is probably not formatted correctly.
5. Line 27: Add comma after "For example"
6. Line 30: Repeated "one example"
7. Line 40: You could probably cite the github package fenics-beat
