### 2024-08-12 finsberg/gotranx#115 @finsberg
> I agree -- this is not a major issue either way. My approach in the past has been to create a package specific `write_model` function, which in your case could be `write_ode(ode_string, filename)`. That way, you can wrap around the Path library's function and get the desired behavior.

OK, I will just keep it as is. There is already a method for saving an ODE to a file (see https://finsberg.github.io/gotranx/docs/api.html#gotranx.ode.ODE.save). In this case I am simply just writing a string to a file and I see no good reason to why I should make a new function to do exactly the same things (except not returning the number characters written).
