### 2024-08-13 finsberg/gotranx#115 @ayush9pandey
OK. So, in principle, I could create an ODE object from the string and then use `save` to save it to a file instead of using the Path library directly?
