### 2024-08-13 0c6fd97
Use ode_from_string in lorentz example instead to saving string to a file
