### 2024-08-13 finsberg/gotranx#115 @finsberg
Yes, there is actually also a function for that, so I updated the example to use that in stead (see https://github.com/finsberg/gotranx/pull/119). Thanks
