### 2024-08-13 5a6f3c4
Add history section on why we decided to do a full rewrite
