+++ joss-paper/paper.bib
@@ -140,3 +140,24 @@
   doi={10.1109/TBME.1978.326270},
   publisher={IEEE}
 }
+
+@article{sundnes2009second,
+  title={A second-order algorithm for solving dynamic cell membrane equations},
+  author={Sundnes, Joakim and Artebrant, Robert and Skavhaug, Ola and Tveito, Aslak},
+  journal={IEEE Transactions on Biomedical Engineering},
+  volume={56},
+  number={10},
+  pages={2546--2548},
+  year={2009},
+  publisher={IEEE}
+}
+
+@software{Finsberg_fenics-beat_2024,
+author = {Finsberg, Henrik},
+doi = {10.5281/zenodo.13323643},
+month = aug,
+title = {{fenics-beat}},
+url = {https://github.com/finsberg/fenics-beat},
+version = {0.1.1},
+year = {2024}
+}
