### 2024-08-17 107f205
Add caption to figure and try with aligned
