### 2024-08-17 finsberg/gotranx#97 @finsberg
I have now added an FAQ section in the README, which hopefully address this issue. A full statement of need is found in the paper. Closing this now, but feel free to reopen if you disagree.
