### 2024-08-17 #7063 @finsberg
@SunnyXu @ayush9pandey I greatly appreciate your comprehensive review. Your insightful comments were extremely beneficial and have significantly improved the quality of the submission. I belive I have now addressed all your issues and also the comments in https://github.com/openjournals/joss-reviews/issues/7063#issuecomment-2282984369

Would be great if you could also tick of the final checks if you agree.

Again, thanks a lot.
