### 2024-08-17 437815c
Fix logo on pypi
