### 2024-08-18 #7063 @SunnyXu
@finsberg
Thanks for the update. I think you have addressed the issues I raised a couple of weeks ago. Currently, both your manuscript and Git/Doc have been greatly improved.
Please note that the two equations between line 65 and Line 66 in the updated manuscript are not aligned exactly.

@HaoZeke
I have updated my review checklist. Please let me know if you have additional requirements for this review. Thanks again for considering me as reviewer. I would be happy to review another work in the future.
