### 2024-08-18 #7063 @ayush9pandey
Thanks for making the changes so quickly @finsberg! Gotranx is a great package and I hope to use it in the near future.

Some very minor typos in the paper:
- Line 22, Line 23, Line 40, space missing when citing.
- Line 73, a comma after "At the time of writing", and then "supports" is probably correct grammatically.
- Labels on the new figure are missing. X-axis is time, Y is the variable.
- The ODE solution figure is a bit too big for my taste, code and figure could have been side by side. But please go ahead and choose the aesthetic that you're most comfortable with. For example, the banner at the top of the README could also be included to create a complete "figure".

@HaoZeke I've updated my review checklist. I don't have any more comments on this JOSS submission. Thanks for inviting to review!
