### 2024-08-18 ed3f60c
Final fixes on paper
