### 2024-08-18 #7063 @finsberg
@ayush9pandey @SunnyXu Thanks, I have now fixed the citations, alignment of the equation and adding the logo in the paper in https://github.com/finsberg/gotranx/pull/121.

I haven't added the labels for the figure. The main reason not to do this is because I don't want to spend another two lines (and four lines in the README) with code that are not really relevant for the software. Also making the figure and code side-by-side would require some markdown magic beyond my expertise (except for creating a new figure with the code and plot side by side which I don't want either). If you have any ideas please let me know. Otherwise, I will keep it as is.

@HaoZeke I have also added the package to Zenodo (see https://zenodo.org/records/13336836). Final version is v1.1.1
