### 2024-08-18 #7063 @ayush9pandey
> I haven't added the labels for the figure. The main reason not to do this is because I don't want to spend another two lines (and four lines in the README) with code that are not really relevant for the software.
Well, there's always the semicolon to have multiple lines on the same line :)

> Also making the figure and code side-by-side would require some markdown magic beyond my expertise (except for creating a new figure with the code and plot side by side which I don't want either). If you have any ideas please let me know.
My idea was a post-edit to make a coherent figure, as is common in publications. But that wouldn't be automated obviously.
