### 2024-08-18 #7063 @HaoZeke
Thanks for the quick turn around on this @finsberg @ayush9pandey and @SunnyXu !  Before finalizing with some of my own comments and handing it off to the EIC I'd like to get the third reviewer's insights as well.

@benlansdell, gentle reminder about this, if it is has fallen off your radar..

P.S. Minor note about figures :: Given that the JOSS version is tagged and archived, I think @ayush9pandey makes a reasonable suggestion, the automated portion can be the rougher figure with an "release task" being to ensure it is post-edited every time.
