### 2024-08-19 9db2942
Attempt to resize figure
