+++ joss-paper/paper.md
@@ -106,7 +106,7 @@ plt.plot(t, y, label="y")
 plt.legend()
 plt.show()
 ```
-![Plot of solution of ODE](../docs/_static/quick_start.png)
+![Plot of solution of ODE](../docs/_static/quick_start.png){ width = 50% }
 
 
 At the time of writing, `gotranx` support code generation to C and Python, with backends for [`numpy`](https://numpy.org) and [`jax`](https://jax.readthedocs.io/). It also supports conversion to and from [CellML](https://www.cellml.org) via [MyoKit](https://myokit.org). For a list of all features and the roadmap, please checkout the [roadmap](https://finsberg.github.io/gotranx/docs/roadmap.html).
