### 2024-08-19 f63e8d0
Put figure into table
