### 2024-08-26 #7063 @benlansdell
Hi all, apologies for being out of touch. I've started looking over the package, and working through the review. I should be done in the next day. So far - this looks like an awesome package, something I wish I had during some of my earlier research projects.
