### 2024-08-28 finsberg/gotranx#126 @benlansdell
General documentation improvements
As part of JOSS review (https://github.com/openjournals/joss-reviews/issues/7063), I noticed the following general things related to the documentation.

It might be good to specify in the README or somewhere else prominent that the python version is required to be below 3.11. (According to `pyproject.toml`. And, indeed, I got an error when trying to run with python 3.12). Would be helpful if the installation section should mention this in particular, as it may require the user create a virtualenv with a different python version.

Though this was brought up (and responded to) in a previous review, I still feel more could be added about the difference between this package and `gotran`. The paper just states "the long term goal is to implement the same features in `gotranx` as found in `gotran` along with additional features." Without saying what these features are, or why gotran is deficient in some way, it doesn't sound like that compelling a pitch.

The page on Vectorized computations (https://finsberg.github.io/gotranx/docs/vectorized_computations.html) feels like it's missing a concluding sentence, at the least. Should the implementation in fenics-beat be used an example for users wanting to implement their own reaction diffusion models? Are there other relevant features/ideas/parts of the code you can direct users to if they want to try implementing reaction-diffusion models with this?
