### 2024-08-28 a2ba6af
Run tests on python3.12
