### 2024-08-28 finsberg/gotranx#128 @benlansdell
How to run performance tests
Can the performance tests (`test_benchmarks.py`) be run outside of CI somehow? They're part of the test suite, but seems like it would require the user setup their own codspeed account, is this right? Possibly a some note can be given about how to setup performance tests. Or the github workflow could file be converted to a script a developer could run (with their own codspeed credentials)?
