### 2024-08-28 finsberg/gotranx#129 @benlansdell
How to get support
The JOSS review guidelines asks us to assess whether the contribution instructions say how to get support, and how to file issues. These are not explicitly stated in `contributing.md`, or anywhere else I could find.

Perhaps a short note could be added making these points explicit.

Also I noticed some very minor typos in `contributing.md`:
- Line 52, For every new feature*
- Line 65, Contributions to the documentation is very welcomed
