### 2024-08-28 #7063 @benlansdell
Ok, have gone through the checklist. All points seem good, with some minor suggestions filed as issues and listed below.

Summary: Gotranx provides a way to convert a high-level specification of an ODE into code, for a variety of languages, that will simulate or solve problems related to the ODE. You can specify different numerical solvers, or just have it work with `scipy.integrate`. This reduces coding required to go from writing down an ODE to simulating or otherwise studying its properties. The flexibility to have it convert to many different languages, or extend it yourself to use a new language, make this a very broadly useful tool. The package provides examples, from computational physiology. It's easy in computational biology, and other fields, for systems of ODEs to get quite intricate, involving many parameters, and possibly dynamics on multiple timescales. Implementing such ODEs can take significant time. This package reduces that time drastically, so should appeal to anyone building and studying such systems.

I filed the following GH issues:
General documentation improvements - https://github.com/finsberg/gotranx/issues/126
How to run performance tests - https://github.com/finsberg/gotranx/issues/128
How to get support - https://github.com/finsberg/gotranx/issues/129

These are all minor points. Overall, this is a very well-written and useful package.
