### 2024-08-28 011b05d
Add cmake as dependency to run tests - this is needed by the compile-c-extension demo
