+++ pyproject.toml
@@ -59,6 +59,7 @@ test = [
     "pytest",
     "pytest-cov",
     "jax",
+    "cmake",
 ]
 benchmark = [
     "pytest-benchmark>=4.0.0",
