### 2024-08-28 d56f254
DeprecatedEnum doesn't work in python3.12 so just get rid of if
