+++ .github/workflows/main.yml
@@ -23,6 +23,7 @@ jobs:
       uses: actions/setup-python@v5
       with:
         python-version: ${{ matrix.python-version }}
+        allow-prereleases: true
     - name: Install gotranx
       run: |
         python -m pip install -e ".[test]"
