### 2024-08-29 53d1ed9
Add more info in vectorized computations
