### 2024-08-29 finsberg/gotranx#126 @finsberg
Hi, thanks for the feedback.

For the issue with python3.12, I have now fixed this here: https://github.com/finsberg/gotranx/pull/127
To ensure that it continues to work on python3.13 I have also now started to test against python3.13 here: https://github.com/finsberg/gotranx/pull/130

For the difference between `gotranx` and `gotran`, I don't know if you saw this section here: https://finsberg.github.io/gotranx/docs/history.html ?
For the features that we plan to implement, we have documented these in the [roadmap](https://finsberg.github.io/gotranx/docs/roadmap.html)

For the page vectorized computations I added a concluding sentence here https://github.com/finsberg/gotranx/pull/131
