### 2024-08-29 b2b5b42
Write about how to run benchmarks locally
