### 2024-08-29 finsberg/gotranx#128 @finsberg
Added more info about this now here: https://github.com/finsberg/gotranx/pull/132
