### 2024-08-29 8114e0b
Remove python version-specific classifiers
