+++ pyproject.toml
@@ -11,10 +11,6 @@ license = {text = "MIT"}
 classifiers = [
     "License :: OSI Approved :: MIT License",
     "Programming Language :: Python :: 3",
-    "Programming Language :: Python :: 3 :: Only",
-    "Programming Language :: Python :: 3.8",
-    "Programming Language :: Python :: 3.9",
-    "Programming Language :: Python :: 3.10",
 ]
 keywords = ["parser", "lark", "gotran", "ode"]
 urls = {Homepage = "https://github.com/finsberg/gotranx"}
