### 2024-08-30 188fc2f
Bump version: 1.1.1 → 1.1.2
