### 2024-08-30 finsberg/gotranx#126 @finsberg
I belive all the issues here are now addressed. Closing this now, but feel free to reopen if you disagree.
