### 2024-08-31 finsberg/gotranx#128 @finsberg
I belive this issue now has been addressed. Feel free to reopen in you disagree.
