### 2024-08-31 #7063 @finsberg
@benlansdell Thanks for the very useful feedback. I really appreciate that your took time to properly review that package. I really think the feedback from your and the other reviewers helped a lot. I have now hopefully addressed all your issues.

@HaoZeke Seems like all reviewers now have completed their review. I have also made a new release (version 1.1.2), and uploaded it to zenodo: https://doi.org/10.5281/zenodo.13621272. Please let me know if there is more you need from me.
