### 2024-09-13 bc1adbd
Add stateful property to Atom and implement a test for singularities
