### 2024-09-13 87c9a27
Add stateful property to Atom and implement a test for singularities
