### 2024-09-16 65943ee
Implmenent code generation for signular odes
