### 2024-09-16 dbc943f
Implmenent code generation for signular odes
