+++ examples/run-python/main.py
@@ -14,6 +14,7 @@ import matplotlib.pyplot as plt
 # We load the model using the {py:func}`load_ode` function
 
 ode = gotranx.load_ode("ORdmm_Land.ode")
+ode = ode.remove_singularities()
 
 # Now we can generate code in python using the `cli` subpackage and the `gotran2py` module`. We will also generate code for the generalized rush larsen scheme
 
