### 2024-09-16 251e045
Fix cornercase with dirac delta - just set it to zero
