### 2024-09-16 1213b8e
Fix cornercase with dirac delta - just set it to zero
