### 2024-09-16 318ff2b
Fix test - we don't handle the case when the limiting value is infinite
