+++ tests/test_python_codegen.py
@@ -742,11 +742,7 @@ def test_codegen_rhs_singular_ode(singular_ode):
         "\n            ),"
         "\n        ),"
         "\n    )"
-        "\n    dx_dt = numpy.where("
-        "\n        numpy.isclose(x, 0),"
-        "\n        numpy.inf * (0.0 if numpy.isclose(b, 0) else numpy.copysign(1, b)),"
-        "\n        b / x,"
-        "\n    )"
+        "\n    dx_dt = b / x"
         "\n    values[0] = dx_dt"
         "\n"
         "\n    return values"
