### 2024-09-16 14f0f21
Make is possible to pass shape
