### 2024-09-16 e39fd7b
Add example with numba
