+++ .pre-commit-config.yaml
@@ -14,7 +14,7 @@ repos:
 
   - repo: https://github.com/astral-sh/ruff-pre-commit
     # Ruff version.
-    rev: 'v0.6.5'
+    rev: 'v0.6.8'
     hooks:
        # Run the linter.
       - id: ruff
