### 2024-10-08 finsberg/gotranx#142 @HaoZeke
[JOSS-Review] Paper related comments
Both #26 and #105 should be completed since they are part of the functionality listed in the paper.
Additionally, I don't see a clear note on what remains to be implemented from `gotran` (and should be in the publication).
It is OK if the tasks cannot be completed but then they must not be noted in the publication (e.g. if `myokit` units are buggy then perhaps not mentioning `myokit` for this version is best).
