### 2024-10-08 finsberg/gotranx#142 @HaoZeke
Part of https://github.com/openjournals/joss-reviews/issues/7063
