### 2024-10-08 finsberg/gotranx#142 @finsberg
@HaoZeke Thanks for your feedback on this.

Regarding #26, this a minor issue that don't really affect the functionality much. The main usage for `gotranx` is converting from CellML using `myokit`, in which the unit conversion works fine, not the other way around.

Regarding #105, we already test the functionality in the documentation (note that the documentation generated here: https://finsberg.github.io/gotranx/examples/new-language/main.html actually runs the code on each push to the repo), so what is missing is adding unit tests for this functionality (we already have integration tests).

Regarding what remains to be implemented from `gotran`, there is a roadmap to be found here: https://finsberg.github.io/gotranx/docs/roadmap.html, but I have no intention to complete this in the near future. This will take years to complete. It is also my understanding that in order to be accepted for a JOSS publication, you don't need to have all the functionality that you want to implement already implemented.

With this in mind I hope you agree that we can close this issue and proceed with accepting the paper for JOSS.
