### 2024-10-11 #7063 @danielskatz
👋 @HaoZeke - please see https://github.com/finsberg/gotranx/issues/142#issuecomment-2400696857
