### 2024-10-14 006d001
[pre-commit.ci] pre-commit autoupdate
updates:
- [github.com/streetsidesoftware/cspell-cli: v8.14.0 → v8.15.1](https://github.com/streetsidesoftware/cspell-cli/compare/v8.14.0...v8.15.1)
