+++ .pre-commit-config.yaml
@@ -31,7 +31,7 @@ repos:
         args: ["--config-file", "pyproject.toml"]
 
   - repo: https://github.com/streetsidesoftware/cspell-cli
-    rev: v8.14.0
+    rev: v8.15.1
     hooks:
       - id: cspell
         files: src/|docs/(.+).md|tests|README.md|examples/(.+).py
