### 2024-10-15 0f03356
Change wording in paper
