### 2024-10-15 72c0e14
Add unit tests for jax
