### 2024-10-15 finsberg/gotranx#142 @finsberg
@HaoZeke I have now changed the wording in the paper and added unit tests for `jax` in https://github.com/finsberg/gotranx/pull/144. Please let me know if you think it needs additional clarification.
