### 2024-10-16 #7063 @HaoZeke
Excellent, @finsberg could you now make a new release, and archive it on Zenodo and let me know as per the [instructions here](https://github.com/openjournals/joss-reviews/issues/7063#issuecomment-2399957159)?
