### 2024-10-16 #7063 @finsberg
@HaoZeke New version (v.1.1.4) is now released. Link to Zendo is https://zenodo.org/records/13938299
