### 2024-10-16 #7063 @HaoZeke
@finsberg could you handle adding the DOI to the `bib` file (and re-doing the release unfortunately).
