+++ joss-paper/paper.bib
@@ -149,7 +149,8 @@
   number={10},
   pages={2546--2548},
   year={2009},
-  publisher={IEEE}
+  publisher={IEEE},
+  doi={10.1109/TBME.2009.2014739}
 }
 
 @software{Finsberg_fenics-beat_2024,
