### 2024-10-16 #7063 @finsberg
@HaoZeke Sorry about the missing DOI, it should be fixed now. I have created a new release (v.1.1.5) and uploaded a new version to Zenodo here: https://zenodo.org/records/13940990
