### 2024-10-16 7aca432
minor changes in paper.md
