+++ joss-paper/paper.md
@@ -35,21 +35,21 @@ The vision for `gotranx` is to implement the same features in `gotranx` as found
 
 # Statement of need
 
-Systems of ordinary Differential Equations (ODEs) are equations of the form
+Systems of ordinary differential equations (ODEs) are equations of the form
 $$
 y'(t) = f(t, y),
 $$
-where $t$ represents time and $y$ is a vector of state variables. These equations are a fundamental concept in mathematics and science, and plays a critical role in various fields such as physics, engineering, and economics.
+where $t$ represents time and $y$ is a vector of state variables. These equations are a fundamental concept in mathematics and science, and play a critical role in various fields such as physics, engineering, and economics.
 
-Solving ODEs can be done analytically in simple cases but for most real world applications one needs to apply numerical methods [@ascher1998computer]. For this there exist a suite of well established softwares such as `scipy` [@virtanen2020scipy] for Python, `Differentialequations.jl` [@rackauckas2017differentialequations] for Julia and Sundials [@hindmarsh2005sundials] for `C` and `C++` .
+Solving ODEs can be done analytically in simple cases but for most real world applications one needs to apply numerical methods [@ascher1998computer]. For this, a suite of well-established software exists, such as `scipy` [@virtanen2020scipy] for Python, `Differentialequations.jl` [@rackauckas2017differentialequations] for Julia, and Sundials [@hindmarsh2005sundials] for `C` and `C++` .
 
-Computational biology is one field where ODEs play an essential role, for example in models of heart cells. The resulting ODEs can, in these cases, be quite involved with many parameters and state variables. For example, one of the more recent models of human heart cells [@tomek2019development] have 112 parameters, 45 state variables and 276 intermediate variables. Solutions of the state equations of these membrane models typically allows for specialized exponential integrators [@rush1978practical] that are implemented as a numerical scheme in `gotranx`.
+Computational biology is one field where ODEs play an essential role, for example in models of heart cells. The resulting ODEs can, in these cases, be quite involved with many parameters and state variables. For example, one of the more recent models of human heart cells [@tomek2019development] has 112 parameters, 45 state variables, and 276 intermediate variables. Solutions of the state equations of these membrane models typically allows for specialized exponential integrators [@rush1978practical] that are implemented as a numerical scheme in `gotranx`.
 
-Furthermore, such models of a single heart cell are typically embedded into a spatial model, such as the Monodomain or Bidomain model [@sundnes2007computing], where each point in a 3D geometry represents a cell. This means that thousands or even millions of such ODEs needs to be coupled and solved in a larger systems. In these cases, traditional solvers are usually unsuitable and custom code are often developed to solve the ODEs. An alternative to this approach is to use an existing framework specialized for these types of problems [@cooper2020chaste; @plank2021opencarp]. However, introducing a big framework might not be ideal if the user wants to do a lot of customizations. With `gotranx` you can easily generate framework independent code that can easily be integrated into most simulation pipelines. One example is [`fenics-beat`](https://github.com/finsberg/fenics-beat) [@Finsberg_fenics-beat_2024] which uses `gotranx` to generate code solving ODEs in a Monodomain model.
[…]
