### 2024-10-16 cc39eaa
minor changes in paper.bib
