+++ joss-paper/paper.bib
@@ -19,7 +19,7 @@
 }
 
 @article{virtanen2020scipy,
-  title={SciPy 1.0: fundamental algorithms for scientific computing in Python},
+  title={SciPy 1.0: fundamental algorithms for scientific computing in {P}ython},
   author={Virtanen, Pauli and Gommers, Ralf and Oliphant, Travis E and Haberland, Matt and Reddy, Tyler and Cournapeau, David and Burovski, Evgeni and Peterson, Pearu and Weckesser, Warren and Bright, Jonathan and others},
   journal={Nature methods},
   volume={17},
@@ -31,7 +31,7 @@
 }
 
 @article{rackauckas2017differentialequations,
-  title={Differentialequations. jl--a performant and feature-rich ecosystem for solving differential equations in julia},
+  title={Differentialequations.jl -- a performant and feature-rich ecosystem for solving differential equations in julia},
   author={Rackauckas, Christopher and Nie, Qing},
   journal={Journal of open research software},
   volume={5},
@@ -108,7 +108,7 @@
 
 
 @article{meurer2017sympy,
-  title={SymPy: symbolic computing in Python},
+  title={SymPy: symbolic computing in {P}ython},
   author={Meurer, Aaron and Smith, Christopher P and Paprocki, Mateusz and {\v{C}}ert{\'\i}k, Ond{\v{r}}ej and Kirpichev, Sergey B and Rocklin, Matthew and Kumar, AMiT and Ivanov, Sergiu and Moore, Jason K and Singh, Sartaj and others},
   journal={PeerJ Computer Science},
   volume={3},
