### 2024-10-16 #7063 @danielskatz
@finsberg - I'm suggesting some small changes in https://github.com/finsberg/gotranx/pull/145 - please merge this, or let me know what you disagree with, then we can continue the process to acceptance and publication
