### 2024-10-17 #7063 @danielskatz
Thanks to @finsberg (Henrik Nicolay Topnes Finsberg) and co-author on your publication!!

JOSS depends on volunteers and we couldn't succeed without you
